Vert.x Apiman Gateway Implementation
This is the initial version of an apiman gateway implemented as a Vert.x module.

The system is composed of several small verticles which run independently and asynchronously; communicating with each
other primarily over a distributed bus (event bus). This is reminiscent of the actor model, and is designed to be
highly scalable, the extent of which which should be fully exposed in the next release:

Whilst this first rendition uses mostly simple in-memory components which don't fully exercise the platform's
distributed scaling potential, the underlying architecture of the gateway is prepared for it; hence, the next release
will implement the necessary distributed components and we will fully expose clustered, multi-verticle configurations.
It is, however, fully functional.

An example configuration file is provided in README.me and src/main/resources/config.json