Removed h2-update step as it's causing build issues
Removed the h2-update build step within the ddl project as it implicitly requires a clean before it will build properly.  This step is part of a work in progress and may be removed in the future, keeping it around until we complete the next phase of the liquibase work.