- bug(gateway-engine-beans) Fixes collision filtration on CaseInsensitiveStringMultiMap #getAllEntries and #getAllValues.
Different bits of code were using #getAllEntries/#getAllValues in different ways. Some required collision filtering, others not. The solution was to overload with additional parameters so that both can be provided.

- spec(gateway-engine-beans) Tests to verify the above

- bug(gateway-engine-beans) Incidentally discovered bug that caused #size to
be inaccurate in certain collision cases. May want to consider deprecating
size as it is likely not quite what people expect and therefore not very
useful.