fix(UI): Workaround incompatiblity between bootstrap tabs and AngularJS 1.6 (and possibly our codebase)
Bootstrap tabs uses anchors with hash fragments; this was triggering AngularJS to route back to the same page, or to the root.
None of the usual workarounds seemed to function as expected such as target="_self" or href="" or using bootstrap data tags instead.

Workaround was to change to button type and reuse styling.

Signed-off-by: Marc Savy <marc@rhymewithgravy.com>