Fix(distro-tomcat): Redirect to Keycloak login page on Tomcat if session invalid.
If the session was not valid there was no redirect to the login page of Tomcat 8.
With this small fix we are able to do a redirect and have a whitelisting of the login CSS files.

This works also if you use keycloak as login provider instead of tomcat. Otherwise also the keycloak login page was not loaded correctly.