Fix example configuration with comma separated list (#650)
* Fix configuration with comma separated list

A comma  separated list work only if the comma is escaped. Because the communs-configuration library (version 1.10) is called with getString(), then that return only the first word.
With the value allowedProtocols=TLSv1.2,TLSv1.1, apiman load only "TLSv1.2".

* Fix configuration with comma separated list

A comma  separated list work only if the comma is escaped. Because the communs-configuration library (version 1.10) is called with getString(), then that return only the first word.
With the value allowedProtocols=TLSv1.2,TLSv1.1, apiman load only "TLSv1.2".

* Fix configuration with comma separated list

A comma  separated list work only if the comma is escaped. Because the communs-configuration library (version 1.10) is called with getString(), then that return only the first word.
With the value allowedProtocols=TLSv1.2,TLSv1.1, apiman load only "TLSv1.2".