fix(gateway-vertx): fix proxy support in plugin resolver
Also adds the ability to specify a proxy via config (i.e. JSON config), 
and SOCKS proxy.

Previously the no proxy list was not handled properly as its separator
was not split up into separate entries, and wildcards were not expanded/matched.

This code has been designed such that it can be lifted and shifted into other
areas of the codebase if we decide to support System properties-defined proxies 
for the Vert.x backend connector also.