ci: retry remote download in case of transient error
Sometimes Maven central responds with an error code. This is usually a transient issue, but it causes the whole build to fail. Instead, by using `-Dmaven.wagon.http.retryHandler.count=3` we can ask the HTTP library to try its request again up to 3 times, giving it a better chance of working.