feat: add Apiman migration assistant
Backstory: We accidentally did not include the Api Schema definition (Swagger, etc) in the Apiman
export file. This means that it vanished when importing into a new version of Apiman. We can't fix
that with the existing tooling very easily because of complexities with Elasticsearch upgrades (the
schema isn't quite right).

So the simplest solution was to create this standalone tool which can export from the old version of
Apiman and then enrich the JSON with the missing information by invoking the Apiman Manager API.

The enriched file can then be imported into the new version of Apiman and the schemas will be
present.

Currently the tool just allows users to enrich pre-2.1.0.Final JSON to include API Schemas but in
future we may need it for other things.

It is presently is done in a very primitive way, and we can refactor it to do chaining of
patches or other stuff we might need if that need occurs.