test: fix improper init of ManagerRestTester JUnit testrunner causing System properties to overwrite one another
As I discovered through a tortuous process, multiple ManagerRestTester JUnit testrunner instances
were being initialised simultaneously. This meant that the code to set environment variables via
`@RestTestSystemProperties`, which is triggered in the constructor, was running in each
ManagerRestTester instance one after another(long before the test code ran). Hence, the environment
variables were overriding each other in surprising and confusing ways (i.e. dependent on
not-entirely-deterministic ordering).

This was compounded by (for whatever reason) CI running the tests in a different order to everyone
else's dev setup, which took a while to notice.

As a bonus, `@RestTestSystemProperties` can now be a reference to a Java resource instead of just a
file. If the file lookup fails, it will try the resource. If nothing is found, a message will be
printed to help diagnose any issues in this area.