fix(manager-api): disable event dispatch during import to prevent event & notification replay
During an import we use the same code paths as a 'real' user interacting with the system,
this causes events to be replayed that would have been generated when the original interaction
occurred. An unwanted side effect of this is that notifications are sent again, which can cause
notifications to be resurrected (with misleading timestamps).

Fixes #1557