chore: add .git-blame-ignore-revs to ignore unimportant format changes
`.git-blame-ignore-revs` allows `git-blame` to ignore certain entries
that are not semantically important. This helps to avoid obscuring
real issues with noise. It will also allow us to apply project-wide
format improvements without worrying about making a mess of the project.

At the moment there's not explicit support in all platforms for this
feature, and a version of git >= 2.23 is required.