feat(manager-api-notifications): add i18n support for notifications
feat(manager-api): add locale to user profile

To better support localised notifications (e.g. via email), we need to store the locale of the 
user we're sending the notification to. Previously we relied on the headers sent in the request from 
the browser to determine preferred language, which only works if there is a request/response 
interaction (i.e. this was not persisted).

Now, on first interaction with the system we take the preferred locale from one of 
(in preferred order):
- Auth token locale (e.g. keycloak)
- Browser language preference
- Default locale

The user's stored locale is preferred over inferred.

The user's locale can be updated through the existing 'update user' REST endpoint.

UI elements will be a separate commit.

Closes: #1624