refactor(ui): improve toc
The table of contents on the 'My Clients' Page needed a better styling.
The toc is now pulled to the right side and the font-size is adjusted for the toc title and the entries.
Futhermore we dont hide the scrollbar anymore if the toc is long enough that it requires a scrollbar.

Fixes: PAS-3291