feat(ui): add tooltips on 'My Clients' page
On the 'My Clients' page only necessary information (client name, API name, plan name) is displayed.
Since it is possible to have multiple APIs with the same name in different organizations it is necessary to make the API organization name visible as well.
Therefore tooltips are added to the client name and the API name are added. They contain all relevant information.

Fixes: PAS-3302