feat(manager-api): if mutable user attributes change in IDM, synchronise user's Apiman profile
If locale, email, or full name changes in IDM (or other security/user info) then we update the profile 
we have stored in Apiman to reflect that. 

Currently we are not updating username to avoid potential breakage. In future we can consider 
allowing this, but will need to add a stable `id` field to users.

Closes #1827