feat(metrics): Drop new RecordMetric in ESMetrics if buffer is full, rather than blocking
If ES goes down the buffer will eventually fill up, and the calling thread will be blocked by the
BlockingQueue. Although this is rare, it can cause the caller to be blocked forever.

This is a particular problem for the Vert.x gateway where it is not acceptable for calling threads
to be blocked for long periods of time.

Fixes: #1884