fix(manager-api-jpa): ensure policies are not left dangling after deletion
If a client / plan / API gets deleted, all attached policies should be deleted as well.

To fix the current behaviour,  where the policies remain in the database after the deletion
of a plan / API / client, a missing `executeUpdate` is added to the deleteAllPolicies function.

To ensure future changes won't break this logic a new delete test is added, as well.

Closes: #1887