fix(ui): restore "requires approval" functionality on devportal tab
Partially reverted changes that got introduced with "fix(ui): show correct discoverability on the plans" (a9552462f86fa2a5d753ffb670d4f8f2c54afb6f).
Changes of the `requiresApproval` were lost because the wrong entity was modified (ApiPlanSummaryBeans instead of ApiPlanBeans).
The summary beans are now only used to display a `planDescription` and the `planName` as these fields are not available in the ApiPlanBean.

Closes: #2071