test(ui): added more cypress tests
We added a bunch of new features to the API Management and the API Developer Portal, therefore we need Cypress E2E tests to ensure that the features work as expected.

The following tests were added:
* Visibility tests for '/home', '/marketplace', '/api-details'
* Infinite scrolling test on '/marketplace'
* Register button test on  '/applications'

In addition to this we added a Cypress section in the readme and provided npm scripts and run configurations to run the tests easily.