test(ui): added cypress test for public endpoint
Since we added a feature that displays the managed endpoint for public APIs on the API details page, we need to test this.

Therefore, we added a cypress test which checks if the managed endpoint of an public API is displayed correctly and the copy btn works as expected.
In addition to this we check if the managed endpoint isn't displayed for private APIs.