fix(ui): fixed position calculation for infinite scrolling
Previously we listened to the scroll event and calculated the position while the scrolling was taking place. This could lead to a buggy behaviour of the infinite scrolling list if the scroll-area was bigger than 400 px.
We added a debounce-timer to debounce the position calculation after the scroll is finished.

Closes: #153