fix(manager-ui-war): force KeyCloakLogoutServlet to use backend URL
In setups where the Keycloak backend URL is different to the frontend URL, Keycloak's Java adapter
currently (incorrectly, AFAICT?) always uses the frontend URL for logging out.

For example, this means it won't actually log out in situations where the internal Docker network
has a different name to the external network, because it attempts to call the frontend URL from
the Apiman backend, which is _inside_ the Docker network, and hence always fails to resolve.

We work around this by forcing the backend URL with a delegate pattern. It's not pretty...