refactor(portal): calculate reasonable API endpoint for Apiman Manager API
The Apiman Developer Portal UI does not know what the Apiman Manager API is unless the user sets
`apiman-manager-ui.api.endpoint`.

As the user is not required to set this essential value, in this commit I add a calculate a
reasonable default based upon the inbound request details. This is identical to the Apiman Manager
UI's approach.

Currently, to avoid too much code churn this is a bit ugly and duplicates some existing code.