feat(metrics-es): allow logging to file with `write-to` option
To facilitate scrape-based metrics, this commit allows Apiman's ES metrics (as JSON) to be written
to a log file. The specific behaviour is controlled via whatever logging framework is used.

It is possible to write to any of: `remote` (ES server), `log`

The configuration provided in this commit calls the output `metrics.log`, but we the user can change
this, as per their preferences. Async logging is used, where possible, to mitigate the impact on
performance, as this is likely to generate a considerable number of log writes.

Fixes: #2100