refactor(manager-rest-api): remove `@RequestBody` annotations
It appears there is a bug in the OpenAPI v3 library we're currently using.

This means that even if you just add a description it no longer produces the schema of the payload,
which is incorrect (it should merge in your description).

Hopefully in the future we can move over to the SmallRye implementation
(which I worked on in the past!) which I think is better.