fix: ensure use of UTC for the time-restricted-access-policy
During the previous deserialization of `java.util.Date` the config was parsed by jackson to local time of the gateway instead of UTC.
Date and joda-time are replaced with java default `OffsetDateTime` which also allows us to be more flexible in future. Old configuration are still compatible.
Test is added back again and fixed.

Closes: #2557